pipeline_retrieval/requirements.txt
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
python-dotenv
requests
